I was messing with things and I made a change to some content.  Even though the content I changed was unrelated to users, I could no longer log in. It was some sort of bug.
I remembered that very recently made a change to the Substance D reference implementation so that the UserToGroup reference type (and other internally-used refernce types) no longer inherit from zope.interface.Interface.  This meant it was became just a plain-old-class.

This object is used as a key in a BTree in the reference map. Unfortunately, because it's just a plain-old-class, the hash computed from it by the BTree machinery is unstable.

http://www.zodb.org/documentation/guide/modules.html#total-ordering-and-persistence

"Another potential problem is mutability: when a key is inserted in a BTree- based structure, it must retain the same order relative to the other keys over time. This is easy to run afoul of if you use mutable objects as keys."

So now references are instances of classes with a total ordering.