Add an 'ISource' interface to '.editable'.
Applications can define an adapter to this interface, allowing the
object to be edited as plain text / RST, or else as HTML.

Provide a 'register_source_adapter' config directive for this adapter.
The directive *also* registers an 'IEditable' adapter for the context,
using an implementation which depends on the semantics of 'ISource'.