Fixed bug by appending rss.xml/index.atom after a url of BlogEntryView.
When user append 'rss.xml' or 'index.atom' after the url of a blog entry,
the app will cause error.

For instance, If there is a blog entry named 'test_entry' and user can
access to '/test_entry' and then user access to '/test_entry/index.atom',
the blog app will cause AttributeError due to try to access 'sdi_title' attribute
of 'Blog Entry' object.

So in this changeset, I applied 'context_type' setting to FeedView to ensure
the views appear on root of the site.